smart_resume_app/requirements.txt
streamlit>=1.24.0
python-docx>=0.8.11
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
keyring>=24.2.0
